Extract MenuItem
